feat: rust number

+++ cpp/include/cpp_number.hpp
@@ -5,7 +5,7 @@
 #include <string>
 
 int cpp_process_number(
-    const int rust_int,
+    const usize rust_int,
     const float rust_float,
     const double rust_double
 );
